fixed serverless for local development
