+++ package.json
@@ -7,6 +7,9 @@
     "node": "^18.15.0"
   },
   "scripts": {
+    "start:debug": "npm run kill:offline && tsc && SLS_DEBUG=* sls offline start",
+    "sqs:invoke-queue-event": "cat tests/resources/queue-event-debug.json | xargs -0 aws --endpoint-url=http://localhost:9324 sqs send-message-batch --queue-url http://localhost:9324/queue/cert-gen-q --entries",
+    "sqs:purge": "aws --endpoint-url=http://localhost:9324 sqs purge-queue --queue-url http://localhost:9324/queue/cert-gen-q",
     "start": "npm run kill:offline && npm run start:offline && serverless invoke local -f main --path tests/resources/queue-event.json",
     "start:offline": "gulp start-serverless",
     "kill:offline": "gulp kill-serverless",
@@ -90,15 +93,16 @@
     "eslint": "^8.57.0",
     "eslint-import-resolver-typescript": "^3.6.1",
     "eslint-plugin-import": "^2.29.1",
-    "gulp": "^3.9.1",
+    "gulp": "^4.0.2",
     "husky": "^3.0.9",
     "jest": "^28.1.3",
     "jest-plugin-context": "^2.9.0",
     "jest-sonar-reporter": "^2.0.0",
     "lambda-tester": "^4.0.1",
     "prettier": "^2.3.2",
-    "serverless": "^3.12.0",
-    "serverless-offline": "^8.5.0",
+    "serverless": "^3.38.0",
+    "serverless-offline": "^13.5.1",
+    "serverless-offline-sqs": "^8.0.0",
     "serverless-plugin-tracing": "^2.0.0",
     "serverless-plugin-typescript": "^2.1.2",
     "sinon": "^7.5.0",
