+++ src/services/S3BucketService.ts
@@ -1,11 +1,8 @@
 import { Inject, Service } from 'typedi';
 import { Readable } from 'stream';
-import AWSXRay from 'aws-xray-sdk';
 import {
   DeleteObjectCommand, DeleteObjectCommandOutput, GetObjectCommand, GetObjectCommandOutput, PutObjectCommand, PutObjectCommandOutput, S3Client,
 } from '@aws-sdk/client-s3';
-import { Configuration } from '../utils/Configuration';
-import { IS3Config } from '../models/IS3Config';
 
 /**
  * Service class for communicating with Simple Storage Service
@@ -15,8 +12,7 @@ class S3BucketService {
   public readonly s3Client: S3Client;
 
   constructor(@Inject() s3Client: S3Client) {
-    const config: IS3Config = Configuration.getInstance().getS3Config();
-    this.s3Client = AWSXRay.captureAWSv3Client(new S3Client({ ...s3Client, ...config }));
+    this.s3Client = s3Client;
   }
 
   /**
